Simple token refactors
