+++ packages/codeeditor/src/index.ts
@@ -1,12 +1,6 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { Token } from '@phosphor/coreutils';
-
-import { IEditorFactoryService } from './factory';
-
-import { IEditorMimeTypeService } from './mimetype';
-
 import '../style/index.css';
 
 export * from './editor';
@@ -14,27 +8,4 @@ export * from './jsoneditor';
 export * from './widget';
 export * from './factory';
 export * from './mimetype';
-
-/* tslint:disable */
-/**
- * Code editor services token.
- */
-export const IEditorServices = new Token<IEditorServices>(
-  '@jupyterlab/codeeditor:IEditorServices'
-);
-/* tslint:enable */
-
-/**
- * Code editor services.
- */
-export interface IEditorServices {
-  /**
-   * The code editor factory.
-   */
-  readonly factoryService: IEditorFactoryService;
-
-  /**
-   * The editor mime type service.
-   */
-  readonly mimeTypeService: IEditorMimeTypeService;
-}
+export * from './tokens';
