+++ packages/console/src/index.ts
@@ -6,5 +6,6 @@ import '../style/index.css';
 export * from './foreign';
 export * from './history';
 export * from './panel';
-export * from './tracker';
 export * from './widget';
+
+export * from './tokens';
