+++ packages/docmanager/src/index.ts
@@ -2,6 +2,7 @@
 // Distributed under the terms of the Modified BSD License.
 
 export * from './dialogs';
+export * from './tokens';
 export * from './manager';
 export * from './savehandler';
 export * from './savingstatus';
