+++ packages/docmanager/src/savingstatus.tsx
@@ -5,7 +5,7 @@ import React from 'react';
 
 import { VDomModel, VDomRenderer } from '@jupyterlab/apputils';
 
-import { IDocumentManager } from './manager';
+import { IDocumentManager } from './tokens';
 
 import { DocumentRegistry } from '@jupyterlab/docregistry';
