+++ packages/documentsearch/src/index.ts
@@ -6,5 +6,6 @@ import '../style/index.css';
 export * from './interfaces';
 export * from './searchinstance';
 export * from './searchproviderregistry';
+export * from './tokens';
 export * from './providers/codemirrorsearchprovider';
 export * from './providers/notebooksearchprovider';
