+++ packages/filebrowser/src/index.ts
@@ -5,7 +5,7 @@ import '../style/index.css';
 
 export * from './browser';
 export * from './crumbs';
-export * from './factory';
+export * from './tokens';
 export * from './listing';
 export * from './model';
 export * from './upload';
