+++ packages/inspector/src/index.ts
@@ -6,3 +6,4 @@ import '../style/index.css';
 export * from './handler';
 export * from './inspector';
 export * from './kernelconnector';
+export * from './tokens';
