+++ packages/rendermime/src/index.ts
@@ -12,4 +12,5 @@ export * from './mimemodel';
 export * from './outputmodel';
 export * from './registry';
 export * from './renderers';
+export * from './tokens';
 export * from './widgets';
