+++ packages/settingeditor/src/tokens.ts
@@ -1,0 +1,23 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+import { IInstanceTracker, MainAreaWidget } from '@jupyterlab/apputils';
+
+import { Token } from '@phosphor/coreutils';
+
+import { SettingEditor } from './settingeditor';
+
+/* tslint:disable */
+/**
+ * The setting editor tracker token.
+ */
+export const ISettingEditorTracker = new Token<ISettingEditorTracker>(
+  '@jupyterlab/settingeditor:ISettingEditorTracker'
+);
+/* tslint:enable */
+
+/**
+ * A class that tracks the setting editor.
+ */
+export interface ISettingEditorTracker
+  extends IInstanceTracker<MainAreaWidget<SettingEditor>> {}
