+++ packages/statusbar/src/index.ts
@@ -7,3 +7,4 @@ export * from './statusbar';
 export * from './style/statusbar';
 export * from './components';
 export * from './defaults';
+export * from './tokens';
