+++ packages/tooltip-extension/src/index.ts
@@ -18,9 +18,9 @@ import {
 
 import { CodeEditor } from '@jupyterlab/codeeditor';
 
-import { IConsoleTracker } from '@jupyterlab/console';
+import { IConsoleTracker } from '@jupyterlab/console/lib/tokens';
 
-import { IEditorTracker } from '@jupyterlab/fileeditor';
+import { IEditorTracker } from '@jupyterlab/fileeditor/lib/tokens';
 
 import { INotebookTracker } from '@jupyterlab/notebook';
